archive/backup_v1_ru_en/requirements.txt
aiogram>=3.4.0
deep-translator>=1.11.4
python-dotenv>=1.0.1
SQLAlchemy>=2.0.29
aiosqlite>=0.20.0
asyncpg>=0.29.0
gTTS>=2.5.0
APScheduler>=3.10.4
pytz>=2024.1
aiohttp>=3.9.0